remove the hardcoded list of blogPosts. We will useState to update that array instead.
instead of just passing down the blogPosts array as the value to the Provider, make it an object. First element in the object will continue to be the array of blogPosts, but now the second element will be a callback function. The call back function will invoke our helper function that calls that setState setter for setBlogPosts but just spreads the new entry from the existing blogPosts array. And creates a new one each time we call addBlogPost. There is nothing to call that function yet. That will be next.