Add in new ShowScreen with basic boiler plate code for now.
Add the ShowScreen to the stackNavigator

destructure out the navigation property from the props in IndexScreen. Wrap the entire Row of the blogPost in TouchableOpacity that navigates to the Show screen. Also add in an additional object that the navigate function will receive and thats the id if the item.