remove the old Add Post button from the index screen the add post functionality is on the CreateScreen now.
make use of the navigation destructured from props to navigate to the Index screen. Instead of just calling that right after the addBlogPost has been called, we will refactor the addBlogPost to take in a callback function. In our case it will be a fat arrow anonymous function that will navigate to Index. So when we onPress now it wil pass in title, content, and this new function.

then in the addBlogPost we will take in this new parameter and call it callback, then after the dispatch action has been called, we will then do callback() and the navigate will occur. We do this patter so we can eventually turn the addBlogPost to an async function that might need to wait for an external api to finish or error out. Then if we error we don't navigate - but for now we don't need this async functionality but it will be possible now.