add new editBlogPost dispatch command and export from the BlogContext so is available in anything that uses the BlogContext provider.
in the EditScreen we did simple refactoring to set the id from navigation.getParam('id') so we don't have to do that each time. We destructure out the editBlogPost function frm the context.

simple refactoring of the state.find to use the new id const we made.

refactor the onSubmit to not call console.log anymore but call our actual editBlogPost function passing in the id of the blogPost to edit and the current title, content state for that blogPost.